choosr-backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
gunicorn==21.2.0
packaging==23.1
psycopg2-binary==2.9.7
sqlparse==0.4.4
typing_extensions==4.7.1
djangorestframework==3.14.0
django-cors-headers==4.2.0
whitenoise==6.5.0
dj-database-url==2.0.0
requests==2.31.0